+++ tests/TestCase/Command/Install/SetPermissionsCommandTest.php
@@ -16,7 +16,6 @@ namespace MeTools\Test\TestCase\Command\Install;
 
 use MeTools\Core\Configure;
 use MeTools\TestSuite\CommandTestCase;
-use Tools\Filesystem;
 
 /**
  * SetPermissionsCommandTest class
@@ -32,7 +31,7 @@ class SetPermissionsCommandTest extends CommandTestCase
         $this->exec('me_tools.set_permissions -v');
         $this->assertExitSuccess();
         foreach (Configure::readFromPlugins('WritableDirs') as $expectedDir) {
-            $this->assertOutputContains('Set permissions on `' . Filesystem::instance()->rtr($expectedDir) . '`');
+            $this->assertOutputContains('Set permissions on `' . rtr($expectedDir) . '`');
         }
     }
 }