+++ tests/TestCase/View/Helper/LibraryHelperTest.php
@@ -38,7 +38,7 @@ class LibraryHelperTest extends HelperTestCase
     {
         parent::setUp();
 
-        Filesystem::instance()->createFile(WWW_ROOT . 'ckeditor' . DS . 'ckeditor.js');
+        Filesystem::createFile(WWW_ROOT . 'ckeditor' . DS . 'ckeditor.js');
     }
 
     /**
@@ -49,7 +49,7 @@ class LibraryHelperTest extends HelperTestCase
     {
         parent::tearDown();
 
-        array_map(fn($dir) => Filesystem::instance()->unlinkRecursive(WWW_ROOT . $dir, '.gitkeep'), ['ckeditor', 'js', 'vendor']);
+        array_map(fn($dir) => Filesystem::unlinkRecursive(WWW_ROOT . $dir, '.gitkeep'), ['ckeditor', 'js', 'vendor']);
     }
 
     /**
@@ -123,7 +123,7 @@ class LibraryHelperTest extends HelperTestCase
         $this->assertHtml($expected, $this->Helper->getView()->fetch('script_bottom'));
 
         //With the jQuery adapter
-        Filesystem::instance()->createFile(WWW_ROOT . 'ckeditor' . DS . 'adapters' . DS . 'jquery.js');
+        Filesystem::createFile(WWW_ROOT . 'ckeditor' . DS . 'adapters' . DS . 'jquery.js');
         $expected = [...$expected, ['script' => ['src' => '/ckeditor/adapters/jquery.js']], '/script'];
         $this->Helper->ckeditor(true);
         $this->assertHtml($expected, $this->Helper->getView()->fetch('script_bottom'));
@@ -136,7 +136,7 @@ class LibraryHelperTest extends HelperTestCase
      */
     public function testCkeditorWithJsFromApp(): void
     {
-        Filesystem::instance()->createFile(WWW_ROOT . 'js' . DS . 'ckeditor_init.js');
+        Filesystem::createFile(WWW_ROOT . 'js' . DS . 'ckeditor_init.js');
 
         $expected = [
             ['script' => ['src' => '/ckeditor/ckeditor.js']],
@@ -155,7 +155,7 @@ class LibraryHelperTest extends HelperTestCase
      */
     public function testCkeditorWithPhpFromApp(): void
     {
-        Filesystem::instance()->createFile(WWW_ROOT . 'js' . DS . 'ckeditor_init.php');
+        Filesystem::createFile(WWW_ROOT . 'js' . DS . 'ckeditor_init.php');
 
         $expected = [
             ['script' => ['src' => '/ckeditor/ckeditor.js']],
@@ -180,7 +180,7 @@ class LibraryHelperTest extends HelperTestCase
         $this->assertSame($expectedJs, $this->Helper->getView()->fetch('script_bottom'));
 
         //With che init file
-        Filesystem::instance()->createFile(WWW_ROOT . 'js' . DS . 'fancybox_init.js');
+        Filesystem::createFile(WWW_ROOT . 'js' . DS . 'fancybox_init.js');
         $expectedJs .= '<script src="/js/fancybox_init.js"></script>';
         $this->Helper->fancybox();
         $this->assertSame($expectedJs, $this->Helper->getView()->fetch('script_bottom'));